requirements.txt
# Core RAG framework
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-voyageai>=0.3.0
langchain-pinecone>=0.2.0
langchain-community>=0.3.0

# PDF Processing
pdfplumber>=0.11.0

# Vector Store
pinecone-client>=5.0.0

# UI
streamlit>=1.40.0

# File Watching
watchdog>=6.0.0

# Environment
python-dotenv>=1.0.0

# Development
pytest>=8.0.0